+++ backend/help.go
@@ -32,7 +32,19 @@ Disabling a cluster prevents creation of new databases or credentials in it and
 immediately disables all existing databases.  
 It will not be possible to renew the lease on a disabled cluster and any active
 lease will be revoked on expiry.
+
+Listing this endpoint lists all active or deleted databases that have been
+registered in the cluster so far.
 `
+
+	helpSynopsisListClusters = `
+List the names of all clusters that have been registered so far
+`
+
+	helpDescriptionListClusters = `
+List clusters path returns a list of all registered clusters, active or
+disabled. Note that the list contains no information about whether the cluster
+is active or not.`
 
 	helpSynopsisClone = `
 Use existing cluster configuration to configure a new cluster.