+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.4
+ - cluster/list: Added the path to list all registered clusters
+ - database/list: Added the path to list all registered databases in a cluster
+
 # 0.1.3
  - cluster/update: New cluster registration generates a random name for management user
  - database/update: New database registration generates a random name for objects owner